Replace line/rect loops
By replacing the writeFillRect, writeFastVLine, and writeFastHLine
for loops calling writePixel with direct calls to fill_rect, vline,
and hline respectively, speed gains of about 660, 54, and 283 times
are achieved, with identical visual result for rectangles of 500x500
pixels, and 500 lines of 500 pixel lengths.

Similar changes to inkplate6.py would probably produce similar gains,
but as I don't have that version of the Inkplate, I can't test that.